Refactor of DVN script
